Converted checklist view from main activity to a fragment
