+++ DroidChecklist/Resources/layout-v21/toolbar.xml
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<android.support.v7.widget.Toolbar xmlns:android="http://schemas.android.com/apk/res/android"
+                                   android:id="@+id/toolbar"
+                                   android:layout_width="match_parent"
+                                   android:layout_height="wrap_content"
+                                   android:elevation="4dp"
+                                   android:minHeight="?attr/actionBarSize"
+                                   android:background="?attr/colorPrimary"
+                                   android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
