+++ DroidChecklist/Resources/layout/Main.axml
@@ -3,30 +3,12 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:minHeight="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
-    <ListView
-        android:id="@+id/list"
+    <include
+        layout="@layout/toolbar"
+        android:id="@+id/toolbar" />
+    <FrameLayout
+        android:id="@+id/content_frame"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_below="@id/toolbar"
-        android:choiceMode="singleChoice" />
-    <android.support.design.widget.CoordinatorLayout
-        android:id="@+id/fab_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/add_fab"
-            android:src="@drawable/ic_add_white_24dp"
-            app:fabSize="normal"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="bottom|end"
-            android:layout_margin="16dp" />
-    </android.support.design.widget.CoordinatorLayout>
+        android:layout_below="@+id/toolbar" />
 </RelativeLayout>
